RESOLVED - task 303: DAOs y servicios para comentarios y like/dislike http://git.fic.udc.es/redmine/issues/show/303
Modificada la lógica de las votaciones.

+++ src/main/java/es/udc/fi/dc/photoalbum/spring/LikeAndDislikeServiceImpl.java
@@ -37,12 +37,10 @@ public class LikeAndDislikeServiceImpl implements LikeAndDislikeService {
 
 	public void voteLike(LikeAndDislike likeAndDislike, User user) {
 		Voted voted = votedDao.get(likeAndDislike.getId(), user.getId());
-		if(voted != null){
-			if(voted.getUserVote().compareTo("LIKE")==0){
-				likeAndDislike.setLike(likeAndDislike.getLike() - 1);
-				likeAndDislikeDao.update(likeAndDislike);
-				votedDao.delete(voted);
-			}
+		if((voted != null)&&(voted.getUserVote().compareTo("LIKE")==0)){
+			likeAndDislike.setLike(likeAndDislike.getLike() - 1);
+			likeAndDislikeDao.update(likeAndDislike);
+			votedDao.delete(voted);
 		}else{
 			voted = new Voted(likeAndDislike, user,"LIKE");
 			votedDao.create(voted);
@@ -53,12 +51,10 @@ public class LikeAndDislikeServiceImpl implements LikeAndDislikeService {
 
 	public void voteDislike(LikeAndDislike likeAndDislike, User user) {
 		Voted voted = votedDao.get(likeAndDislike.getId(), user.getId());
-		if(voted != null){
-			if(voted.getUserVote().compareTo("DISLIKE")==0){
-				likeAndDislike.setDislike(likeAndDislike.getDislike() - 1);
-				likeAndDislikeDao.update(likeAndDislike);
-				votedDao.delete(voted);
-			}
+		if((voted != null)&&(voted.getUserVote().compareTo("DISLIKE")==0)){
+			likeAndDislike.setDislike(likeAndDislike.getDislike() - 1);
+			likeAndDislikeDao.update(likeAndDislike);
+			votedDao.delete(voted);
 		}else{
 			voted = new Voted(likeAndDislike, user,"DISLIKE");
 			votedDao.create(voted);
